+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Bugfix) Fix and document action timeouts
 - (Feature) Propagate sidecars' ports to a member's service
 - (Debug Package) Initial commit
+- (Feature) Detach PVC from deployment in Ordered indexing method
 
 ## [1.2.16](https://github.com/arangodb/kube-arangodb/tree/1.2.16) (2022-09-14)
 - (Feature) Add ArangoDeployment ServerGroupStatus